+++ web-frame/src/app-shared/config/replacements/available-feature-modules.ts
