+++ web-frame/src/app-shared/translate/index.ts
@@ -10,9 +10,9 @@
 
 export * from './translate.module';
 
-export * from './defaults/available-languages-metadata';
-export * from './defaults/ngx-translate-config';
 export * from './interfaces/language-metadata.interface';
 export * from './interfaces/translation-phrase.interface';
 export * from './other/default-language.token';
+export * from './replacements/available-languages-metadata';
+export * from './replacements/ngx-translate-config';
 export * from './services/translation.service';
