+++ web-frame/src/app-shared/translate/replacements/available-languages-metadata.ts
