+++ web-frame/src/app-shared/translate/replacements/ngx-translate-config.ts
